+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.38.0...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- add errors for directive options
 - add errors for invalid comment directives
 - add function name to summary report header